Cleaning tool inspector

+++ CodeLibraryProject/Assets/HANDINS-NILS/Voxels/ProgressBar.cs
@@ -13,7 +13,7 @@ public class ProgressBar : EditorWindow
     {
         EditorUtility.DisplayProgressBar("Voxel creation progress",
             $"Creating voxels... {pCurrentVoxelIndex}/{MaxVoxelIndex}",
-            pCurrentVoxelIndex / MaxVoxelIndex);
+            (float)pCurrentVoxelIndex / MaxVoxelIndex);
 
         if (pCurrentVoxelIndex < MaxVoxelIndex) return;
         if (!HasFinishedProcess) return;
@@ -26,7 +26,7 @@ public class ProgressBar : EditorWindow
     {
         EditorUtility.DisplayProgressBar("Voxel collision detection progress",
             $"Detecting if voxels are traversable... {pCurrentVoxelIndex}/{MaxVoxelIndex}",
-            pCurrentVoxelIndex / MaxVoxelIndex);
+            (float)pCurrentVoxelIndex / MaxVoxelIndex);
 
         if (pCurrentVoxelIndex < MaxVoxelIndex) return;
         if (!HasFinishedProcess) return;
@@ -39,7 +39,7 @@ public class ProgressBar : EditorWindow
     {
         EditorUtility.DisplayProgressBar("Voxel neighbour detection progress",
             $"Detecting what voxels are neighbours of current voxel... {pCurrentVoxelIndex}/{MaxVoxelIndex}",
-            pCurrentVoxelIndex / MaxVoxelIndex);
+            (float)pCurrentVoxelIndex / MaxVoxelIndex);
 
         if (pCurrentVoxelIndex < MaxVoxelIndex) return;
         if (!HasFinishedProcess) return;
